REFACT: move menu to own .py file
